fix: add Css Vars story
